test(ci): group chaos integration targets
