+++ tests/grouped/chaos/chaos_migration_sigkill_resume.rs
@@ -28,6 +28,7 @@
 #![cfg(unix)]
 
 #[allow(dead_code)]
+#[path = "../../support/mod.rs"]
 mod support;
 
 use std::path::PathBuf;
