+++ tests/grouped/chaos/chaos_promote_refused_when_lease_held.rs
@@ -9,6 +9,7 @@
 //! generic 409.
 
 #[allow(dead_code)]
+#[path = "../../support/mod.rs"]
 mod support;
 
 use reddb::replication::lease::{LeaseError, LeaseStore};
