+++ tests/grouped/chaos/chaos_wal_segment_missing.rs
@@ -1,20 +1,18 @@
-//! Chaos test: WAL segment payload corruption (PLAN.md Phase 2.4 / 8 slice).
+//! Chaos test: missing middle WAL segment (PLAN.md Phase 11.3 / 8 slice).
 //!
-//! Archives three WAL segments with valid sha256 sidecars, then
-//! tampers segment 2's recorded sha256 to a value that doesn't match
-//! the on-disk payload. Restore must fail closed before applying any
-//! record from the corrupted segment.
+//! Archives three segments, then deletes segment 2's payload + sidecar.
+//! Restore enumerates segments 1 and 3 (segment 2 is gone from the
+//! bucket); segment 3's `prev_hash` points to segment 2's sha256, not
+//! segment 1's, so the chain check fires.
 
 use reddb::storage::backend::LocalBackend;
-use reddb::storage::wal::{
-    archive_change_records, load_wal_segment_manifest, publish_snapshot_manifest,
-    publish_wal_segment_manifest, PointInTimeRecovery,
-};
+use reddb::storage::wal::{archive_change_records, publish_snapshot_manifest, PointInTimeRecovery};
 use reddb::storage::RedDB;
-use reddb_file::SnapshotManifest;
+use reddb_file::{wal_segment_manifest_key, SnapshotManifest};
 use std::sync::Arc;
 
 #[allow(dead_code)]
+#[path = "../../support/mod.rs"]
 mod support;
 
 fn temp_dir(prefix: &str) -> support::TempDataDir {
@@ -26,8 +24,8 @@ fn record(lsn: u64, payload: &[u8]) -> reddb::replication::cdc::ChangeRecord {
 }
 
 #[test]
-fn restore_fails_closed_on_segment_sha256_mismatch() {
-    let work = temp_dir("sha-corrupt");
+fn restore_fails_closed_on_missing_middle_segment() {
+    let work = temp_dir("missing");
     let snapshot_dir = work.join("snapshots");
     let wal_dir = work.join("wal");
     let restore_path = work.join("restore").join("data.rdb");
@@ -68,13 +66,13 @@ fn restore_fails_closed_on_segment_sha256_mismatch() {
         metas.push(m);
     }
 
-    // Tamper segment 2's recorded sha256 — payload bytes are intact,
-    // but the sidecar lies about their digest.
-    let mut sidecar2 = load_wal_segment_manifest(&LocalBackend, &metas[1].key)
-        .unwrap()
-        .expect("segment 2 sidecar");
-    sidecar2.sha256 = Some("ff".repeat(32));
-    publish_wal_segment_manifest(&LocalBackend, &sidecar2).unwrap();
+    // Delete segment 2's payload AND sidecar — gone from the bucket
+    // entirely. The LocalBackend stores keys as filesystem paths, so
+    // `metas[1].key` is the on-disk path.
+    let seg2_path = std::path::Path::new(&metas[1].key);
+    let seg2_sidecar = wal_segment_manifest_key(&metas[1].key);
+    let _ = std::fs::remove_file(seg2_path);
+    let _ = std::fs::remove_file(&seg2_sidecar);
 
     let recovery = PointInTimeRecovery::new(
         Arc::new(LocalBackend),
@@ -83,10 +81,10 @@ fn restore_fails_closed_on_segment_sha256_mismatch() {
     );
     let err = recovery
         .restore_to(0, &restore_path)
-        .expect_err("sha256 mismatch must fail closed");
+        .expect_err("missing middle segment must fail closed");
     let msg = err.to_string().to_lowercase();
     assert!(
-        msg.contains("integrity") || msg.contains("sha256"),
-        "error must mention integrity/sha256; got: {msg}"
+        msg.contains("chain"),
+        "error must mention chain; got: {msg}"
     );
 }
